+++ python/aistore/pytorch/aisio.py
@@ -1,6 +1,6 @@
 """
 AIS IO Datapipe
-Copyright (c) 2022-2023, NVIDIA CORPORATION. All rights reserved.
+Copyright (c) 2022-2025, NVIDIA CORPORATION. All rights reserved.
 """
 
 from typing import Iterator, Tuple, List
@@ -12,6 +12,7 @@ from torchdata.datapipes.iter import IterDataPipe
 from torchdata.datapipes.utils import StreamWrapper
 
 from aistore.sdk.ais_source import AISSource
+from aistore.sdk.etl import ETLConfig
 
 try:
     from aistore.sdk import Client
@@ -146,7 +147,7 @@ class AISFileLoaderIterDataPipe(IterDataPipe[Tuple[str, StreamWrapper]]):
             yield url, StreamWrapper(
                 self.client.bucket(bck_name=bck_name, provider=provider)
                 .object(obj_name=obj_name)
-                .get_reader(etl_name=self.etl_name)
+                .get_reader(etl=ETLConfig(name=self.etl_name))
                 .raw()
             )
 
@@ -156,7 +157,9 @@ class AISFileLoaderIterDataPipe(IterDataPipe[Tuple[str, StreamWrapper]]):
 
 @functional_datapipe("ais_list_sources")
 class AISSourceLister(IterDataPipe[str]):
-    def __init__(self, ais_sources: List[AISSource], prefix="", etl_name=None):
+    def __init__(
+        self, ais_sources: List[AISSource], prefix="", etl: ETLConfig = None
+    ) -> None:
         """
         Iterable DataPipe over the full URLs for each of the provided AIS source object types
 
@@ -164,17 +167,17 @@ class AISSourceLister(IterDataPipe[str]):
             ais_sources (List[AISSource]): List of types implementing the AISSource interface: Bucket, ObjectGroup,
              Object, etc.
             prefix (str, optional): Filter results to only include objects with names starting with this prefix
-            etl_name (str, optional): Pre-existing ETL on AIS to apply to all selected objects on the cluster side
+            etl (ETLConfig, optional): Pre-existing ETL on AIS to apply to all selected objects on the cluster side
         """
         _assert_aistore()
         self.sources = ais_sources
         self.prefix = prefix
-        self.etl_name = etl_name
+        self.etl = etl
 
     def __getitem__(self, index) -> T_co:
         raise NotImplementedError
 
     def __iter__(self) -> Iterator[T_co]:
         for source in self.sources:
-            for url in source.list_urls(prefix=self.prefix, etl_name=self.etl_name):
+            for url in source.list_urls(prefix=self.prefix, etl=self.etl):
                 yield url