+++ python/aistore/sdk/etl/__init__.py
@@ -1,0 +1,6 @@
+#
+# ETL Module
+# Copyright (c) 2025, NVIDIA CORPORATION. All rights reserved.
+#
+
+from aistore.sdk.etl.etl_config import ETLConfig